blackbox: append-only flight log in flash bank 2
The enabler for the measurements a bench cannot make -- inertia and
drag_k need the airframe free to move, which means flight, which means
the wire is not available. logger::putc busy-waits on USART6 TXE inside a
critical_section, so a sample logged as it is taken stalls the 8 kHz loop
for milliseconds, and there is nobody holding a cable anyway.

Bank 2, for two independent reasons, both load-bearing:

  - DFU writes only bank 1, so a log SURVIVES A REFLASH. The whole
    workflow rests on that: fly a `blackbox` build, then flash a
    `blackbox-dump` build and read back the flight you just did. Without
    it you would need a way off the board before reflashing, which is the
    hard part of every blackbox.
  - The H7 cannot read a bank while programming it. Code executes from
    bank 1, so writing bank 2 does not stall the instruction fetch.

memory.x now splits accordingly: FLASH 1024K (bank 1, image is ~205K),
BLACKBOX 896K, CONFIG 128K. Verified nothing lands past 0x08031840.

Append-only with no wrap. H7 flash programs 32-byte words and a word may
be written ONCE between erases -- writing it twice, even with the same
value, corrupts its ECC and the read fails. Wrapping would need a
multi-second erase mid-flight, and losing the start of a flight to keep
the end is the wrong trade when the interesting transients are at the
start. 28672 records, 143 s at the default 200 Hz.

Erased flash reads 0xFF and PlantSample::encode zeroes its reserved tail,
so "written" and "never written" are distinguishable with no index and no
CRC. The append point is found by BINARY search, not a scan: the log is
append-only so written records are a strict prefix, which is exactly what
a binary search needs. 15 reads at boot instead of 28672.

PlantSample::t_ms widened u16 -> u32. 65.5 s was ample for a four-second
bench capture and is a short hop in flight, and a wrapped timestamp fits
a plausible, wrong time constant without ever looking broken.

Staging: a 256-record channel, and the control loop uses try_send so a
slow flash can never stall it. Drops are counted and reported per second
rather than silently lost -- a log with invisible holes fits a wrong
model as confidently as a complete one. The drain lives in persist_task
because that task already owns the flash handle, and two owners of a
blocking flash driver cannot be made safe by being careful. The two
cannot collide: cal saves happen disarmed, blackbox writes armed.

KNOWN RISK, UNMEASURED. control_loop is awaited on the same single
embassy executor as persist_task, and embassy's executor is cooperative,
so every flash program blocks the 8 kHz loop for the duration of the
write -- tens of us to ~100 us against a 125 us period, ~200 times a
second. Whether that matters is a question about real hardware timing and
I have no hardware, so rather than guess, the control loop now reports
its MEASURED rate once a second ("control loop NNNN Hz measured"). If it
sags with `blackbox` on, the fix is a lower rate or an interrupt
executor for the control loop. Check this before trusting a flight log.

Both features are off by default and the flight build is byte-identical
in size without them.

Tests: 352 pass (10 new, covering the region geometry against memory.x,
blank-vs-zero, and the binary search's boundary and read count). All five
build configurations compile; Renode quick tier passes.

+++ src/plant_fit.rs
@@ -112,7 +112,7 @@ pub fn find_steps(
             let mut j = i + 1;
             while j < samples.len()
                 && (samples[j].cmd_f32(motor) - now).abs() < MIN_STEP
-                && samples[j].t_ms.wrapping_sub(t0) <= max_window_ms
+                && samples[j].t_ms.wrapping_sub(t0) <= max_window_ms as u32
             {
                 j += 1;
             }
@@ -380,16 +380,16 @@ mod tests {
         e_to: f32,
         quantise: bool,
     ) -> std::vec::Vec<PlantSample> {
-        let dt_ms = (1000.0 / hz) as u16;
+        let dt_ms = (1000.0 / hz) as u32;
         let mut out = std::vec::Vec::new();
-        let mut t: u16 = 0;
+        let mut t: u32 = 0;
         // Steady before the step.
-        while t < hold_ms {
+        while t < hold_ms as u32 {
             out.push(mk(t, 0.10, e_from, quantise));
             t += dt_ms;
         }
         let t_step = t;
-        while t < hold_ms * 2 {
+        while t < hold_ms as u32 * 2 {
             let dt = (t - t_step) as f32 * 1e-3;
             let e = e_to + (e_from - e_to) * (-dt / tau_omega).exp();
             out.push(mk(t, 0.20, e, quantise));
@@ -398,7 +398,7 @@ mod tests {
         out
     }
 
-    fn mk(t_ms: u16, cmd: f32, erpm: f32, quantise: bool) -> PlantSample {
+    fn mk(t_ms: u32, cmd: f32, erpm: f32, quantise: bool) -> PlantSample {
         let period = 60_000_000.0 / erpm;
         let p = if quantise { period as u16 } else { period.round() as u16 };
         PlantSample {
